requirements.txt
asgiref==3.7.2
Django==5.0.1
django-filter==23.5
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
pytz==2023.3.post1
PyYAML==6.0.1
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
uritemplate==4.1.1
